pyproject.toml
[project]
name = "generative-and-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "ollama>=0.5.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "wikipedia>=1.4.0",
]
